Search should be working now

+++ admin.py
@@ -81,7 +81,7 @@ class LoadTestDataHandler(database.webapp2.RequestHandler):
           li.put()
           database.logging.info("Creating test data user_id=%s", li.user_id)
           for j in range(0, 10):
-            item = database.Item(title="item" + str((i * 10) + j), description="This is a description of the item. It may contain various amounts of information.", price='%.2f' % float((i * 10) + j), expiration_date=database.datetime.date.today() + database.datetime.timedelta(weeks=4), is_active=True, deactivated=False, created_by_id=li.user_id)
+            item = database.Item(title="item" + str((i * 10) + j), description="This is a description of the item. It may contain various amounts of information.", price=float('%.2f' % float((i * 10) + j)), expiration_date=database.datetime.date.today() + database.datetime.timedelta(weeks=4), is_active=True, deactivated=False, created_by_id=li.user_id)
             if (len(item.description) > 40):
               item.summary = item.description[:40].rstrip() + "..."
             else:
